// src/app/models/subject.model.ts
export interface Subject {
  userId: string;
  subjectId: string;
  materia: string;
  anio: string;
  cuatrimestre: string;
  estado: string;
  horario: string;
  fechaPrimerParcial: string;
  notaPrimerParcial: string;
  fechaSegundoParcial: string;
  notaSegundoParcial: string;
  fechaRecuperatorio: string;
  notaRecuperatorio: string;
  fechaFinal: string;
  notaFinal: string;
  observaciones: string;
}